The FastScroller is not working properly.
If you have a large number of contacts in the last section of contacts list,
when you scroll the list, the fast scroll bar will work improperly, it will scroll back.

The reason is that when we scroll to the last section, the next position of the next section is -1 but not the count of the contacts.

Change-Id: I058b3e6c88e8aaeabbfed291296e02dfd775018e